requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
psycopg2-binary==2.9.6
Werkzeug==2.3.7
email_validator==2.0.0
Flask-Migrate==4.0.4
gunicorn==21.2.0
python-json-logger==2.0.7  # Add this for better JSON logging